rename folder store and api
