+++ packages/client/src/store/rootReduser.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import { authApi } from 'store/services/auth/authApi';
-import userReducer from 'store/services/auth/userInfoSlice';
-import { rtkApi } from 'store/services/gameApi/rtk';
+import { authApi } from 'api/auth/authApi';
+import userReducer from 'store/userInfoSlice';
+import { rtkApi } from 'api/gameApi/rtk';
 
 export const rootReducer = combineReducers({
 	[authApi.reducerPath]: authApi.reducer,
